+++ src/main/java/kuehlfrank/backend/model/Unit.java
@@ -19,6 +19,6 @@ public class Unit {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long unitId;
+	private Integer unitId;
 	private String label;
 }
