+++ src/main/java/kuehlfrank/backend/repositories/IngredientAlternativeNameRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface IngredientAlternativeNameRepository extends CrudRepository<IngredientAlternativeName, Long>{
+public interface IngredientAlternativeNameRepository extends CrudRepository<IngredientAlternativeName, Integer>{
 
     @Query("SELECT ing FROM IngredientAlternativeName ing WHERE ing.name = :ingredientAlternativeName")
     Optional<IngredientAlternativeName> findByName(String ingredientAlternativeName);
