+++ src/main/java/kuehlfrank/backend/repositories/InventoryEntryRepository.java
@@ -8,11 +8,11 @@ import org.springframework.data.repository.CrudRepository;
 import kuehlfrank.backend.model.Ingredient;
 import kuehlfrank.backend.model.InventoryEntry;
 
-public interface InventoryEntryRepository extends CrudRepository<InventoryEntry, Long> {
+public interface InventoryEntryRepository extends CrudRepository<InventoryEntry, Integer> {
 
 	@Query("SELECT inve FROM InventoryEntry inve WHERE inve.ingredient.ingredientId = :ingredientId AND inve.unit.unitId = :unitId")
-	Optional<InventoryEntry> findByIngredientAndUnitId(Long ingredientId, Long unitId);
+	Optional<InventoryEntry> findByIngredientAndUnitId(Integer ingredientId, Integer unitId);
 
 	@Query("SELECT invEntry.ingredient from InventoryEntry invEntry where invEntry.inventory.inventoryId = :inventoryEntryId")
-	Optional<Ingredient> findByInventoryEntryId(Long inventoryEntryId);
+	Optional<Ingredient> findByInventoryEntryId(Integer inventoryEntryId);
 }
