+++ src/main/java/kuehlfrank/backend/repositories/UnitRepository.java
@@ -3,6 +3,6 @@ package kuehlfrank.backend.repositories;
 import kuehlfrank.backend.model.Unit;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UnitRepository extends CrudRepository<Unit, Long>{
+public interface UnitRepository extends CrudRepository<Unit, Integer>{
 
 }
